Modified order.js and restructured api.py.
